huge test refactor
pull object/expected-filter generator into a test/support/ module
pull assert-equals into test/support/ module
change where "npm test" points
